BF: array images return array if OK float type
The `get_fdata` method should return the contained array if the array is
the correct (matching) floating point type.  This was not happening
because of I used the `astype` method, which, by default, does a copy.
Fix and test.

+++ nibabel/dataobj_images.py
@@ -344,7 +344,7 @@ class DataobjImage(FileBasedImage):
         if self._fdata_cache is not None:
             if self._fdata_cache.dtype.type == dtype.type:
                 return self._fdata_cache
-        data = np.asanyarray(self._dataobj).astype(dtype)
+        data = np.asanyarray(self._dataobj).astype(dtype, copy=False)
         if caching == 'fill':
             self._fdata_cache = data
         return data
